+++ file/file.go
@@ -210,9 +210,10 @@ type GetFileByIDResponse struct {
 }
 
 type getSharedFilesResponse struct {
-	Files     []*GetFileByIDResponse `json:"files"`
-	PageNum   int64                  `json:"pageNum"`
-	ItemCount int64                  `json:"itemCount"`
+	Files     *filesResponse `json:"files"`
+	PageNum   int64          `json:"pageNum"`
+	ItemCount int64          `json:"itemCount"`
+	ErrMsg    string         `json:"errMsg,omitempty"`
 }
 
 type partialFile struct {
@@ -226,6 +227,11 @@ type partialFile struct {
 	CreatedAt   int64   `json:"createdAt,omitempty"`
 	UpdatedAt   int64   `json:"updatedAt,omitempty"`
 	Float       bool    `json:"float,omitempty"`
+}
+
+type filesResponse struct {
+	Successful []*GetFileByIDResponse `json:"successful"`
+	Failed     []string               `json:"failed"`
 }
 
 type updateFilesRequest struct {
@@ -501,15 +507,16 @@ func (r *Router) GetSharedFiles(c *gin.Context, queryAppID string) {
 		return
 	}
 
-	// Go over the permissions and get the files metadata related to them
-	files := make([]*GetFileByIDResponse, 0, len(permissions.GetPermissions()))
+	// Make an empty slice of files
+	filesSuccesful := make([]*GetFileByIDResponse, 0, len(permissions.GetPermissions()))
+	filesFailed := make([]string, 0, len(permissions.GetPermissions()))
+
 	for _, permission := range permissions.GetPermissions() {
 		file, err := r.fileClient().GetFileByID(c.Request.Context(), &fpb.GetByFileByIDRequest{Id: permission.GetFileID()})
 		if err != nil {
-			httpStatusCode := gwruntime.HTTPStatusFromCode(status.Code(err))
-			loggermiddleware.LogError(r.logger, c.AbortWithError(httpStatusCode, err))
-
-			return
+			loggermiddleware.LogError(r.logger, fmt.Errorf("failed fetching file %v: %v", permission.GetFileID(), err))
+			filesFailed = append(filesFailed, permission.GetFileID())
+			continue
 		}
 
 		// Filter files which belong to the requesting user.
@@ -521,17 +528,24 @@ func (r *Router) GetSharedFiles(c *gin.Context, queryAppID string) {
 				Role:    permission.GetRole(),
 				Creator: permission.GetCreator(),
 			}
-			files = append(
-				files,
+			filesSuccesful = append(
+				filesSuccesful,
 				CreateGetFileResponse(file, permission.GetRole().String(), userPermission),
 			)
 		}
 	}
 
+	var errMsg string
+	if len(filesFailed) > 0 {
+		errMsg = "file not found"
+	}
+
+	files := &filesResponse{Successful: filesSuccesful, Failed: filesFailed}
 	sharedFilesResponse := &getSharedFilesResponse{
 		Files:     files,
 		PageNum:   permissions.PageNum,
 		ItemCount: permissions.ItemCount,
+		ErrMsg:    errMsg,
 	}
 
 	c.JSON(http.StatusOK, sharedFilesResponse)